# sample_components/co_swagger_0.0.7.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.7",
    "title": "Customer Onboarding At RAK Bank",
    "termsOfService": "http://rakbank.ae/api/terms/",
    "contact": {
      "email": "api-support@rakbank.ae"
    },
    "description": "A typical customer onboarding process involves multiple activities. It starts with evaluating the customer's requirement, followed by due-diligence, completion of legal aspects, account opening and making it operational."
  },
  "host": "rakbank.ae",
  "basePath": "/api/customer-onboarding/v1",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/banks/RAK/usertypes/sme/prospects/": {
      "post": {
        "summary": "post a prospect for SME accounts and services ",
        "operationId": "createSMEProspect",
        "tags": [
          "Onboard SME Customer"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "SmeProspect object to create new prospect",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SmeProspect"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully saved prospect details",
            "schema": {
              "$ref": "#/definitions/SmeProspectResponse"
            }
          },
          "400": {
            "description": "400 is the generic client-side error status. Errors can be like malformed request syntax, invalid request message parameters etc. ",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "500 is the generic REST API error response.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/banks/RAK/usertypes/{userType}/prospects/search": {
      "post": {
        "summary": "search prospect by usertype, mobileNo, leadNumber",
        "operationId": "searchProspect",
        "tags": [
          "Search Prospect"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userType",
            "in": "path",
            "required": true,
            "description": "usertype e.g. sme, retail, corporate",
            "type": "string",
            "enum": [
              "sme",
              "retail",
              "corporate"
            ]
          },
          {
            "in": "body",
            "name": "body",
            "description": "search prospect",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SearchProspectRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/SearchProspectResponse"
            }
          },
          "400": {
            "description": "Invalid prospectId",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "500 is the generic REST API error response.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/banks/RAK/usertypes/sme/prospects/{prospectId}": {
      "get": {
        "summary": "Get prospect by prospectId",
        "operationId": "getSMEProspectById",
        "tags": [
          "Get SME Prospect"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "prospectId",
            "in": "path",
            "required": true,
            "description": "The id of the prospect to retrieve",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/SmeProspect"
            }
          },
          "400": {
            "description": "Invalid prospectId",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "500 is the generic REST API error response.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "summary": "Update an existing prospect by prospectId",
        "operationId": "updateSMEProspect",
        "tags": [
          "Update SME Prospect"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "prospectId",
            "in": "path",
            "required": true,
            "description": "The id of the prospect to update",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "SmeProspect object to update existing prospect",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SmeProspect"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully updated prospect details",
            "schema": {
              "$ref": "#/definitions/SmeProspectResponse"
            }
          },
          "400": {
            "description": "400 is the generic client-side error status. Errors can be like malformed request syntax, invalid request message parameters etc. ",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "500 is the generic REST API error response.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/banks/RAK/prospects/{prospectId}/documents/{documentId}": {
      "get": {
        "summary": "get document by documentId",
        "operationId": "getDocumentById",
        "tags": [
          "Get Document by Id"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "prospectId",
            "in": "path",
            "required": true,
            "description": "The id of the prospect",
            "type": "string"
          },
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "description": "The id of the document",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "return document",
            "schema": {
              "$ref": "#/definitions/ProspectDocument"
            }
          },
          "400": {
            "description": "400 is the generic client-side error status. Errors can be like malformed request syntax, invalid request message parameters etc.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "500 is the generic REST API error response.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "PreScreeningInfo": {
      "type": "object",
      "properties": {
        "statusOverAll": {
          "type": "string",
          "enum": [
            "proceed",
            "stop"
          ]
        },
        "screeningResults": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "screeningType": {
                "type": "string",
                "description": ""
              },
              "screeningStatus": {
                "type": "string",
                "description": ""
              },
              "screeningFlag": {
                "type": "string",
                "description": ""
              },
              "screeningNotes": {
                "type": "string",
                "description": ""
              },
              "screeningReason": {
                "type": "string",
                "description": ""
              },
              "reasonNotes": {
                "type": "string",
                "description": ""
              },
              "reasonCode": {
                "type": "string",
                "description": ""
              },
              "reasonStatus": {
                "type": "string",
                "description": ""
              }
            }
          }
        }
      }
    },
...
    "SearchProspectRequest": {
      "type": "object",
      "properties": {
        "fname": {
          "type": "string",
          "maxLength": 50
        },
        "countryCode": {
          "type": "string",
          "description": "country code",
          "pattern": "^[0-9]*$",
          "maxLength": 3
        },
        "mobileNo": {
          "type": "string",
          "description": "mobile number",
          "pattern": "^[0-9]*$",
          "maxLength": 15
        },
        "leadNumber": {
          "type": "string",
          "description": "RAK track lead generated by the system e.g. ABCD",
          "maxLength": 9999
        },
        "tradeLicenseNo": {
          "type": "string",
          "pattern": "^[a-zA-Z0-9]*$",
          "maxLength": 20
        },
        "email": {
          "type": "string",
          "description": "email Id of the applicant",
          "format": "email",
          "maxLength": 50
        },
        "eidNumber": {
          "type": "string",
          "maxLength": 15,
          "pattern": "^[0-9]*$"
        }
      }
    },
    "SmeProspectResponse": {
      "type": "object",
      "properties": {
        "accountInfo": {
          "type": "array",
          "items": {
            "properties": {
              "accountNo": {
                "type": "string"
              },
              "currency": {
                "type": "string"
              }
            }
          }
        },
        "prospectId": {
          "type": "string",
          "maxLength": 50
        },
        "_links": {
          "type": "object",
          "properties": {
            "self": {
              "type": "object",
              "properties": {
                "href": {
                  "type": "string"
                }
              }
            }
          }
        },
        "preScreening": {
          "$ref": "#/definitions/PreScreeningInfo"
        }
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "errorType": {
          "type": "string",
          "enum": [
            "FieldsValidation",
            "OTP",
            "Other"
          ]
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "fieldPath": {
                "type": "string"
              },
              "errorCode": {
                "type": "string"
              },
              "errorType": {
                "type": "string",
                "enum": [
                  "mandatory",
                  "invalid"
                ]
              },
              "message": {
                "type": "string"
              },
              "developerText": {
                "type": "string",
                "description": "pattern, maxlength, min, max etc used for invalid fields "
              }
            }
          }
        }
      }
    }
  }
}